treewide: rename CONFIG_SPL_I2C_SUPPORT to CONFIG_SPL_I2C
This is basically the same change as provided by Masahiro Yamada [1], but
with the Kconfig elements added.

We conventionally named SPL options in CONFIG_SPL_*_SUPPORT format,
but the suffix _SUPPORT did not add any value.  Besides, it prevents
us from exploiting the CONFIG_IS_ENABLED() macro.

If we move the option to defconfig files (that are often re-synced),
the renaming would be more conflict-happy, so I want to do this now.

[1] http://lists.denx.de/pipermail/u-boot/2016-September/266311.html

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>